added a type check in __eq__ method of ReturnPrint and restructured execute
So that if an error is thrown during execution (which returns a string
signaling that) it doesn’t try to compare a string to something that
might not be a string. Type check is because apparently this was not
enough and it was still trying to compare that string to a ReturnPrint
and crashing.